// backend/functions/src/old/functions/connect/imported/interfaces.ts
export interface Price {
    /**
     * Whether the price can be used for new purchases.
     */
    active: boolean;
    currency: string;
    unit_amount: number;
    /**
     * A brief description of the price.
     */
    description: string | null;
    /**
     * One of `one_time` or `recurring` depending on whether the price is for a one-time purchase or a recurring
     * (subscription) purchase.
     */
    type: 'one_time' | 'recurring';
    /**
     * The frequency at which a subscription is billed. One of `day`, `week`, `month` or `year`.
     */
    interval: 'day' | 'month' | 'week' | 'year' | null;
    /**
     * The number of intervals (specified in the `interval` attribute) between subscription billings. For example,
     * `interval=month` and `interval_count=3` bills every 3 months.
     */
    interval_count: number | null;
    /**
     * Default number of trial days when subscribing a customer to this price using
     * [`trial_from_plan=true`](https://stripe.com/docs/api#create_subscription-trial_from_plan).
     */
    trial_period_days: number | null;

    /**
     * Any additional properties
     */
    [propName: string]: any;
}

export interface Product {
    /**
     * Whether the product is currently available for purchase.
     */
    active: boolean;
    /**
     * The product's name, meant to be displayable to the customer. Whenever this product is sold via a subscription,
     * name will show up on associated invoice line item descriptions.
     */
    name: string;
    /**
     * The product's description, meant to be displayable to the customer. Use this field to optionally store a long
     * form explanation of the product being sold for your own rendering purposes.
     */
    description: string | null;
    /**
     * A list of up to 8 URLs of images for this product, meant to be displayable to the customer.
     */
    images: Array<string>;
    /**
     * A list of Prices for this billing product.
     */
    prices?: Array<Price>;

    stripeAccountId: string;

    chatBotConfig: FirebaseFirestore.DocumentReference<FirebaseFirestore.DocumentData>;

    default_price: string;

    /**
     * Any additional properties
     */
    [propName: string]: any;
}